requirements.txt
azure-core #==1.31.0
azure-datalake-store #==0.0.53
azure-identity #==1.18.0
azure-storage-blob #==12.23.0
docker #==7.1.0
dvc #==3.55.2
dvc-azure #==3.1.0
fastapi #==0.115.0
fastapi-cli #==0.0.5
joblib #==1.4.2
#jwt
kaggle #==1.6.17
matplotlib #==3.9.2
mlflow #==2.16.2
mlflow-skinny #==2.16.2
pandas # ==2.2.3
passlib #==1.7.4
pillow #==10.4.0
pydantic #==2.9.2
pydantic_core #==2.23.4
PyJWT>=2.0.0
pyjwt
python-dotenv #==1.0.1
requests #==2.32.3
scikit-learn #==1.5.2
streamlit #==1.38.0
torch #==2.4.1
tqdm #==4.66.5
transformers #==4.45.1
uvicorn #==0.30.6
opencv-python #==4.10.0.84
scikit-image #==0.24.0
nibabel #==5.2.1
tensorflow #==2.17.0
tensorflow-io-gcs-filesystem #==0.37.1
graphviz #==0.20.3
python-multipart
bcrypt==3.1.7
evidently==0.4.38